tester.hpp: make 'double == float' comparison use double internally, like 'float == double' comparison

+++ tests/tester.hpp
@@ -141,9 +141,9 @@ namespace tester
 	template<>
 	bool equal_t<value_t<double>, value_t<float>>::operator()() const
 	{
-		const float epsilon = 0.000001;
-		const float upper   = rhs() + epsilon;
-		const float lower   = rhs() - epsilon;
+		const double epsilon = 0.000001;
+		const double upper   = rhs() + epsilon;
+		const double lower   = rhs() - epsilon;
 		return lower <= lhs() && lhs() <= upper;
 	}
